+++ CHANGELOG.md
@@ -12,7 +12,8 @@ The format of the release notes follows [Keep a Changelog](https://keepachangelo
   download sessions ([#16])
 - Method `clean` to clean the product folder recursively from any files not belonging to the inventory
   ([#16])
-- Method `ignore` to exclude database files in the inventory ([#17])
+- Methods `ignore!` and `unignore!` to exclude database files in the inventory and include 
+  them again ([#17])
 
 ### Changed